Mass Flow Controller Software
This software controls a mass flow controller through COM port
communication. Made for Windows XP, it is written in C# and  uses .NET
framework 4.0. AsyncBridge is used to support the asyncronous functions
that are not available in .NET framework 4.0. The Mass Flow Controller
is connected to the PC through a USB-to-RS485 converter. The
communication protocol used to send commands to the MFC are for a Brooks
(Celerity) MFC, however most MFCs should follow the same protocols.